started implementing enables and waveformview
started implementing enable switches for each parameter as well as the pre/post waveform but need to turn the wave form into eq waterfall  or add it. Also only turned off knobs not the signal processing not sure if this will help much on processing need to look at the numbers.